Moved InMemoryMultipartFile out from migration code to prevent deletion
