+++ src/main/resources/application.yaml
@@ -80,7 +80,7 @@ spring:
 
   servlet:
     multipart:
-      max-file-size: 2MB
+      max-file-size: 4MB
   data:
     web:
       pageable:
@@ -121,4 +121,4 @@ fact:
     audit:
       retention-days: ${AUDIT_RETENTION_DAYS:365}
     photo:
-      max-width: ${PHOTO_MAX_WIDTH:640}
+      max-width: ${PHOTO_MAX_WIDTH:1024}
