+++ src/test/java/uk/gov/hmcts/reform/fact/data/api/migration/model/InMemoryMultipartFileTest.java
@@ -11,6 +11,8 @@ import java.nio.file.Path;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import uk.gov.hmcts.reform.fact.data.api.models.InMemoryMultipartFile;
+
 class InMemoryMultipartFileTest {
 
     @TempDir
